+++ src/main/java/fr/openmc/core/features/adminshop/AdminShopManager.java
@@ -11,124 +11,29 @@ import fr.openmc.core.utils.messages.MessagesManager;
 import fr.openmc.core.utils.messages.Prefix;
 import lombok.Getter;
 import net.kyori.adventure.text.Component;
-import org.bukkit.ChatColor;
 import org.bukkit.Material;
-import org.bukkit.configuration.file.FileConfiguration;
-import org.bukkit.configuration.file.YamlConfiguration;
 import org.bukkit.entity.Player;
 import org.bukkit.inventory.ItemStack;
 import org.bukkit.inventory.PlayerInventory;
 
-import java.io.File;
-import java.io.IOException;
 import java.text.DecimalFormat;
 import java.util.*;
-import java.util.logging.Level;
 
 public class AdminShopManager {
     private final OMCPlugin plugin;
-    private FileConfiguration config;
-    private final File configFile;
     public final Map<String, ShopCategory> categories = new HashMap<>();
     public final Map<String, Map<String, ShopItem>> items = new HashMap<>();
     public final Map<UUID, String> currentCategory = new HashMap<>();
     public final DecimalFormat priceFormat = new DecimalFormat("#,##0.00");
+    private final AdminShopYAML adminShopYAML;
 
     @Getter private static AdminShopManager instance;
 
     public AdminShopManager(OMCPlugin plugin) {
         instance = this;
         this.plugin = plugin;
-        this.configFile = new File(plugin.getDataFolder() + "/data", "adminshop.yml");
-        loadConfig();
-    }
-
-    private void loadConfig() {
-        if (!configFile.exists()) plugin.saveResource("data/adminshop.yml", false);
-        config = YamlConfiguration.loadConfiguration(configFile);
-        loadCategories();
-        loadItems();
-    }
-
-    private void loadCategories() {
-        categories.clear();
-        List<Map<?, ?>> categoryList = config.getMapList("category");
-
-        for (Map<?, ?> categoryMap : categoryList) {
-            for (Map.Entry<?, ?> entry : categoryMap.entrySet()) {
-                String key = entry.getKey().toString();
-                Map<?, ?> section = (Map<?, ?>) entry.getValue();
-                categories.put(key, new ShopCategory(
-                        key,
-                        ChatColor.translateAlternateColorCodes('&', section.get("name").toString()),
-                        Material.valueOf(section.get("material").toString()),
-                        (int) section.get("position")
-                ));
-            }
-        }
-    }
-
-    private void loadItems() {
-        items.clear();
-
-        for (String categoryId : categories.keySet()) {
-            List<Map<?, ?>> itemList = config.getMapList(categoryId);
-            Map<String, ShopItem> categoryItems = new HashMap<>();
-
-            for (Map<?, ?> itemMap : itemList) {
-                for (Map.Entry<?, ?> entry : itemMap.entrySet()) {
-                    String itemKey = entry.getKey().toString();
-                    Map<?, ?> itemSection = (Map<?, ?>) entry.getValue();
-
-                    String name = ChatColor.translateAlternateColorCodes('&', itemSection.get("name").toString());
-                    int slot = (int) itemSection.get("slot");
-                    Material material = Material.valueOf(itemKey);
-
-                    Map<?, ?> prices = (Map<?, ?>) itemSection.get("price");
-                    Map<?, ?> initial = (Map<?, ?>) prices.get("initial");
-                    Map<?, ?> actual = (Map<?, ?>) prices.get("actual");
-
-                    categoryItems.put(itemKey, new ShopItem(
-                            itemKey, name, material, slot,
-                            Double.parseDouble(initial.get("sell").toString()),
-                            Double.parseDouble(initial.get("buy").toString()),
-                            Double.parseDouble(actual.get("sell").toString()),
-                            Double.parseDouble(actual.get("buy").toString())
-                    ));
-                }
-            }
-
-            if (!categoryItems.isEmpty()) items.put(categoryId, categoryItems);
-        }
-    }
-
-    public void saveConfig() {
-        for (var entry : items.entrySet()) {
-            String categoryId = entry.getKey();
-            List<Map<String, Object>> itemList = new ArrayList<>();
-
-            for (var itemEntry : entry.getValue().entrySet()) {
-                ShopItem item = itemEntry.getValue();
-                Map<String, Object> itemData = Map.of(
-                        "name", item.getName(),
-                        "slot", item.getSlot(),
-                        "price", Map.of(
-                                "initial", Map.of("sell", item.getInitialSellPrice(), "buy", item.getInitialBuyPrice()),
-                                "actual", Map.of("sell", item.getActualSellPrice(), "buy", item.getActualBuyPrice())
-                        )
-                );
-
-                itemList.add(Map.of(itemEntry.getKey(), itemData));
-            }
-
-            config.set(categoryId, itemList);
-        }
-
-        try {
-            config.save(configFile);
-        } catch (IOException e) {
-            plugin.getLogger().log(Level.SEVERE, "Could not save shop config", e);
-        }
+        this.adminShopYAML = new AdminShopYAML(plugin, this);
+        this.adminShopYAML.loadConfig();
     }
 
     public void openBuyConfirmMenu(Player player, String categoryId, String itemId, Menu previousMenu) {
@@ -142,7 +47,7 @@ public class AdminShopManager {
         ShopItem item = getItemSafe(player, categoryId, itemId);
         if (item == null) return;
 
-        if (!playerHasItem(player, item.getMaterial(), 1)) {
+        if (playerHasItem(player, item.getMaterial(), 1)) {
             sendError(player, "Vous n'avez pas cet item dans votre inventaire !");
             return;
         }
@@ -183,7 +88,7 @@ public class AdminShopManager {
             return;
         }
 
-        if (!playerHasItem(player, item.getMaterial(), amount)) {
+        if (playerHasItem(player, item.getMaterial(), amount)) {
             sendError(player, "Vous n'avez pas assez de " + item.getName() + " à vendre !");
             return;
         }
@@ -207,7 +112,7 @@ public class AdminShopManager {
         item.setActualSellPrice(Math.max(newSell, item.getInitialSellPrice() * 0.5));
         item.setActualBuyPrice(Math.max(newBuy, item.getInitialBuyPrice() * 0.5));
 
-        saveConfig();
+        this.adminShopYAML.saveConfig();
     }
 
     private int checkInventorySpace(Player player, Material itemToAdd, int amountToAdd) {
@@ -247,9 +152,9 @@ public class AdminShopManager {
     private boolean playerHasItem(Player player, Material material, int amount) {
         int count = 0;
         for (ItemStack item : player.getInventory().getContents()) {
-            if (item != null && item.getType() == material && (count += item.getAmount()) >= amount) return true;
+            if (item != null && item.getType() == material && (count += item.getAmount()) >= amount) return false;
         }
-        return false;
+        return true;
     }
 
     private void removeItems(Player player, Material material, int amount) {
@@ -300,7 +205,7 @@ public class AdminShopManager {
     }
 
     public String formatPrice(double value) {
-        return "$" + priceFormat.format(value);
+        return priceFormat.format(value) + " " + EconomyManager.getEconomyIcon();
     }
 
     public void openMainMenu(Player player) {
