+++ src/main/java/fr/openmc/core/features/adminshop/AdminShopYAML.java
@@ -1,0 +1,121 @@
+package fr.openmc.core.features.adminshop;
+
+import fr.openmc.core.OMCPlugin;
+import org.bukkit.ChatColor;
+import org.bukkit.Material;
+import org.bukkit.configuration.file.FileConfiguration;
+import org.bukkit.configuration.file.YamlConfiguration;
+import org.jetbrains.annotations.NotNull;
+
+import java.io.File;
+import java.io.IOException;
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.logging.Level;
+
+public class AdminShopYAML {
+    private final OMCPlugin plugin;
+    private FileConfiguration config;
+    private final File configFile;
+    private final AdminShopManager shopManager;
+
+    public AdminShopYAML(OMCPlugin plugin, AdminShopManager shopManager) {
+        this.plugin = plugin;
+        this.configFile = new File(plugin.getDataFolder() + "/data", "adminshop.yml");
+        this.shopManager = shopManager;
+    }
+
+    public void loadConfig() {
+        if (!configFile.exists()) plugin.saveResource("data/adminshop.yml", false);
+        config = YamlConfiguration.loadConfiguration(configFile);
+        loadCategories();
+        loadItems();
+    }
+
+    private void loadCategories() {
+        shopManager.categories.clear();
+        List<Map<?, ?>> categoryList = config.getMapList("category");
+
+        for (Map<?, ?> categoryMap : categoryList) {
+            for (Map.Entry<?, ?> entry : categoryMap.entrySet()) {
+                String key = entry.getKey().toString();
+                Map<?, ?> section = (Map<?, ?>) entry.getValue();
+                shopManager.categories.put(key, new ShopCategory(
+                        key,
+                        ChatColor.translateAlternateColorCodes('&', section.get("name").toString()),
+                        Material.valueOf(section.get("material").toString()),
+                        (int) section.get("position")
+                ));
+            }
+        }
+    }
+
+    private void loadItems() {
+        shopManager.items.clear();
+
+        for (String categoryId : shopManager.categories.keySet()) {
+            List<Map<?, ?>> itemList = config.getMapList(categoryId);
+            Map<String, ShopItem> categoryItems = new HashMap<>();
+
+            for (Map<?, ?> itemMap : itemList) {
+                for (Map.Entry<?, ?> entry : itemMap.entrySet()) {
+                    String itemKey = entry.getKey().toString();
+                    Map<?, ?> itemSection = (Map<?, ?>) entry.getValue();
+
+                    String name = ChatColor.translateAlternateColorCodes('&', itemSection.get("name").toString());
+                    int slot = (int) itemSection.get("slot");
+                    Material material = Material.valueOf(itemKey);
+
+                    Map<?, ?> prices = (Map<?, ?>) itemSection.get("price");
+                    Map<?, ?> initial = (Map<?, ?>) prices.get("initial");
+                    Map<?, ?> actual = (Map<?, ?>) prices.get("actual");
+
+                    categoryItems.put(itemKey, new ShopItem(
+                            itemKey, name, material, slot,
+                            Double.parseDouble(initial.get("sell").toString()),
+                            Double.parseDouble(initial.get("buy").toString()),
+                            Double.parseDouble(actual.get("sell").toString()),
+                            Double.parseDouble(actual.get("buy").toString())
+                    ));
+                }
+            }
+
+            if (!categoryItems.isEmpty()) shopManager.items.put(categoryId, categoryItems);
+        }
+    }
+
+    public void saveConfig() {
+        for (var entry : shopManager.items.entrySet()) {
+            String categoryId = entry.getKey();
+            List<Map<String, Object>> itemList = new ArrayList<>();
+
+            for (var itemEntry : entry.getValue().entrySet()) {
+                Map<String, Object> itemData = convertShopItemToMap(itemEntry);
+
+                itemList.add(Map.of(itemEntry.getKey(), itemData));
+            }
+
+            config.set(categoryId, itemList);
+        }
+
+        try {
+            config.save(configFile);
+        } catch (IOException e) {
+            plugin.getLogger().log(Level.SEVERE, "Could not save shop config", e);
+        }
+    }
+
+    private static @NotNull Map<String, Object> convertShopItemToMap(Map.Entry<String, ShopItem> itemEntry) {
+        ShopItem item = itemEntry.getValue();
+        return Map.of(
+                "name", item.getName(),
+                "slot", item.getSlot(),
+                "price", Map.of(
+                        "initial", Map.of("sell", item.getInitialSellPrice(), "buy", item.getInitialBuyPrice()),
+                        "actual", Map.of("sell", item.getActualSellPrice(), "buy", item.getActualBuyPrice())
+                )
+        );
+    }
+}
